requirements.txt
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
python-dotenv==1.0.0
pydantic==1.10.13
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
newspaper3k==0.2.8
yfinance==0.2.33
google-search-results==2.4.2